requirements.txt
anthropic>=0.40.0
requests>=2.32.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.2
python-whois>=0.9.4
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
